+++ src/Generic/AppHelp.cs
@@ -1,0 +1,8 @@
+﻿using System.Reflection;
+using System.Linq;
+
+class AppHelp
+{
+    public static string ProductName { get; } = (Assembly.GetEntryAssembly()?.GetCustomAttributes(
+        typeof(AssemblyProductAttribute)).First() as AssemblyProductAttribute)?.Product ?? "";
+}