+++ src/Misc.cs
@@ -1,60 +1,28 @@
-﻿using Microsoft.Win32;
-using System;
+﻿using System;
 
-public class Item
+[Serializable()]
+public class AppSettings
+{
+    public bool RawView { get; set; } = true;
+    public bool ShowCompactSummary { get; set; } = true;
+    public bool WordWrap { get; set; }
+    public bool CenterScreen { get; set; } = true;
+    public bool FormatEncoded { get; set; } = true;
+    
+    public double FontSize { get; set; } = 14;
+    public int WindowWidth { get; set; } = 750;
+    public int WindowHeight { get; set; } = 550;
+    public int ColumnPadding { get; set; } = 30;
+    
+    public string FontName { get; set; } = "Consolas";
+    public string Exclude { get; set; } = "";
+    public string FileTypes { get; set; } = "avi mp4 mkv mov webm mpg vob ts m2v m2ts mts m4v wmv flv asf avc 264 h264 hevc 265 h265 mp2 mp3 m4a mpa aac ogg opus mka dts dtsma dtshr dtshd ac3 eac3 thd wav w64 flac sub sup idx ass aas srt png jpg jpeg gif bmp";
+}
+
+public class MediaInfoParameter
 {
     public string Name { get; set; } = "";
     public string Value { get; set; } = "";
     public string Group { get; set; } = "";
     public bool IsComplete { get; set; }
-}
-
-public class RegHelp
-{
-    public static void SetValue(string path, string? name, object value)
-    {
-        using (RegistryKey? regKey = GetRootKey(path)?.CreateSubKey(path.Substring(5), RegistryKeyPermissionCheck.ReadWriteSubTree))
-            regKey?.SetValue(name, value);
-    }
-
-    public static string GetString(string path, string? name, string defaultValue = "")
-    {
-        object? value = GetValue(path, name, defaultValue);
-        return !(value is string) ? defaultValue : value.ToString() ?? "";
-    }
-
-    public static int GetInt(string path, string? name, int defaultValue = 0)
-    {
-        object? val = GetValue(path, name, defaultValue);
-        return !(val is int) ? defaultValue : (int)val;
-    }
-
-    public static object? GetValue(string path, string? name, object? defaultValue = null)
-    {
-        using (RegistryKey? regKey = GetRootKey(path)?.OpenSubKey(path.Substring(5)))
-            return regKey == null ? defaultValue : regKey.GetValue(name, defaultValue);
-    }
-
-    public static void RemoveKey(string path)
-    {
-        GetRootKey(path)?.DeleteSubKeyTree(path.Substring(5), false);
-    }
-
-    public static void RemoveValue(string path, string name)
-    {
-        using (RegistryKey? regKey = GetRootKey(path)?.OpenSubKey(path.Substring(5), true))
-            regKey?.DeleteValue(name, false);
-    }
-
-    static RegistryKey? GetRootKey(string path)
-    {
-        switch (path.Substring(0, 4))
-        {
-            case "HKLM": return Registry.LocalMachine;
-            case "HKCU": return Registry.CurrentUser;
-            case "HKCR": return Registry.ClassesRoot;
-        }
-
-        return null;
-    }
 }