+++ Changelog.md
@@ -1,5 +1,16 @@
 
+### 5.0
+
 - changelog added to repo
 - if no file is open message is shown
 - file can be opened from menu: Open File...
+- file can be saved from menu: Save File...
 - exclude setting added to hide defined parameters
+- encoding presentation can be disabled
+- compact summary can be disabled
+- file types for setup can be customized
+- column width can be customized
+- new settings dialog with proper UI controls and greatly improved usability
+- some dialogs use task dialog instead of msg box
+- about dialog added
+- settings directory can either be AppData, Portable or Custom